Split maths-related utilities to their own cog
