+++ bjokeybot/cogs/__init__.py
@@ -1,7 +1,7 @@
 from bjokeybot.cogs.fun import FunCog
 from bjokeybot.cogs.stats import StatsCog
 from bjokeybot.cogs.tts import TTSCog
-from bjokeybot.cogs.util import UtilityCog
+from bjokeybot.cogs.util import GeneralUtilityCog, MathUtilityCog
 from bjokeybot.cogs.chat_admin import ChatAdminCog
 
-all_cogs = (FunCog, StatsCog, TTSCog, UtilityCog, ChatAdminCog)
+all_cogs = (FunCog, StatsCog, TTSCog, GeneralUtilityCog, MathUtilityCog, ChatAdminCog)
